fix: move serialport mock to /mock subpath export
