+++ packages/serial/src/index.ts
@@ -5,7 +5,6 @@ export * from "./message/INodeQuery";
 export * from "./message/Message";
 export * from "./message/SuccessIndicator";
 export * from "./MessageHeaders";
-export * from "./MockSerialPort";
 export * from "./parsers/BootloaderParsers";
 export * from "./parsers/SerialAPIParser";
 export * from "./ZWaveSerialPort";
