+++ packages/serial/src/index_mock.ts
@@ -1,2 +1,3 @@
+export * from "./MockSerialPort";
 export * from "./SerialPortBindingMock";
 export * from "./SerialPortMock";
