+++ packages/zwave-js/src/lib/test/driver/handleUnsolicited.test.ts
@@ -1,5 +1,6 @@
 import { BasicCCValues } from "@zwave-js/cc/BasicCC";
-import { MessageHeaders, MockSerialPort } from "@zwave-js/serial";
+import { MessageHeaders } from "@zwave-js/serial";
+import type { MockSerialPort } from "@zwave-js/serial/mock";
 import { createThrowingMap, ThrowingMap } from "@zwave-js/shared";
 import { wait } from "alcalzone-shared/async";
 import type { Driver } from "../../driver/Driver";
