+++ packages/zwave-js/src/lib/test/driver/sendDataFailThrow.test.ts
@@ -1,5 +1,6 @@
 import { BasicCCSet } from "@zwave-js/cc";
-import { MessageHeaders, MockSerialPort } from "@zwave-js/serial";
+import { MessageHeaders } from "@zwave-js/serial";
+import type { MockSerialPort } from "@zwave-js/serial/mock";
 import type { ThrowingMap } from "@zwave-js/shared";
 import { wait } from "alcalzone-shared/async";
 import type { Driver } from "../../driver/Driver";
